Fixed: Publish Date in Admin Does not Save in Correct Format
There&#039;s a problem with the Publish Date box when saving or updating a post in Admin. When selecting a date from the date picker for locale English Australia, the date is saved in American format. For example, selecting 1 June 2020 from the date picker the post publish date is saved as January 6 2020. To save as 1 June 2020 you have to manually enter it in the box in American format - 06/01/2020.