Fixed issue with wrong comment date on non en_US locales
Report from the customer:
The dates the extension retrieves for the comments is in incorrect. I have debugged the code, and found that at the point the code converts a date-time string to a timestamp, the timestap that comes back is incorrect. For example, a date of 2020-04-11 comes back as a date in 2016. October 10, 2016 to be precise. I also checked when I add a comment today in the M2 site, even then the creation date of the comment is incorrect: Mar 12, 2032 17:04. In the database the field is: 2020-09-26 17:04:00 which is correct. The M2 backend does show the correct date though, but the extension show the date in 2032. For now I will remove the date on the comments as it shows a bit silly having comments from the future :-) I am using the &quot;magefan&quot;-comments as defined in the M2 backend